removed max_people from courses table. added some 'funcions' usability (delete function, some checks). added views in DB (kursy_aktualne, kursy_zakonczone) added some checks to courses. removes some warnings from php (datetime and timezone) fixed Messages from postgress. Now www site shows errors. improved view of messages.

+++ CreateDB/CREATE/TRIGGERS/FUNKCJE.sql
@@ -1,9 +1,20 @@
 CREATE OR REPLACE FUNCTION funkcje_usun_fun() RETURNS TRIGGER AS'
 	DECLARE
+		temp INTEGER;
 		w funkcje_ekipa_kurs%rowtype;
 	BEGIN
-		DELETE FROM funkcje_ekipa_kurs WHERE id_funkcje = OLD.id;
-		RETURN OLD;
+		SELECT id FROM (
+			SELECT id_kursy as id 
+			FROM funkcje_ekipa_kurs fek 
+			WHERE fek.id_funkcje = old.id) as t 
+		INTO temp
+		WHERE id in (SELECT id FROM kursy_aktualne); 
+		IF FOUND THEN
+			RAISE EXCEPTION ''Nie możesz usunąć funkcji % . Aktualnie odbywają się kursy które jej używają! Usuń najpierw wszystkie odwołania do tej funkcji w aktualnych kursach. <br/> Do informacji: Możesz usunąć funkcje jeśli odwołuje się tylko do minionych kursów. Zniknie wtedy jakikolwiek ślad po niej.'', OLD.id;
+		ELSE 
+			DELETE FROM funkcje_ekipa_kurs WHERE id_funkcje = OLD.id;
+			RETURN OLD;
+		END IF;
 	END
 ' LANGUAGE 'plpgsql';
